backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
pytesseract==0.3.10
spacy==3.7.4
deep-translator==1.11.4
python-dotenv==1.0.1
Werkzeug==3.0.1
Pillow==10.2.0
PyMuPDF==1.23.26
transformers==4.38.2
torch==2.2.1
scikit-learn==1.4.1.post1
SpeechRecognition==3.10.1
PyPDF2==3.0.1
pdf2image==1.17.0
numpy==1.26.4
nltk==3.8.1
pyttsx3==2.90
gTTS==2.5.1
pandas==2.2.1
